Refactor handling of workdir, file removals in signing classes
Move construction of a signing object to after the creation
of the temp directory, so we can stash it in the object
and avoid passing it through to the methods.

And since all file removals should be relative to the
working directory, refactor the utils functions to take
that path separately, reducing risk that we'll accidentally
delete files elsewhere due to coding errors.

+++ digsigserver/utils.py
@@ -6,7 +6,10 @@ from sanic import request
 from sanic.log import logger
 
 
-def remove_file(path: str, check=False):
+def remove_file(workdir: str, name: str, check: bool = False):
+    if os.path.isabs(name):
+        raise ValueError("{}: absolute path not allowed".format(name))
+    path = os.path.join(workdir, name)
     if os.path.isdir(path):
         shutil.rmtree(path, ignore_errors=not check)
     else:
@@ -18,9 +21,9 @@ def remove_file(path: str, check=False):
             pass
 
 
-def remove_files(paths, check=False):
+def remove_files(workdir: str, paths: list, check: bool = False):
     for path in paths:
-        remove_file(path, check)
+        remove_file(workdir, path, check)
 
 
 def extract_files(workdir: str, f: request.File) -> bool:
